requirements.txt
pandas==2.2.2
openpyxl==3.1.5
python-dateutil==2.9.0.post0
unidecode==1.3.8
pyyaml==6.0.2
streamlit==1.35.0
